backend/app/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
requests>=2.31.0
python-multipart>=0.0.6
sentence-transformers>=2.2.0
Pillow>=10.0.0
openai>=1.10.0
